Refactor RTDE implementation. rename file back to original. remove custom rtde protocol
